Phase 0: Extract shared lmdb-stream module, consolidate duplicated code
- Created scripts/lmdb-stream.ts: shared streaming helpers (toHex,
  formatUtxoEntry, formatGenericEntry, createLmdbStream, getSnapshotInfo,
  CORS_HEADERS) extracted from duplicated code in both server files.
- Refactored mithril-stream-server.ts to import from lmdb-stream.ts.
- Refactored gerolamo-explorer.ts to import from lmdb-stream.ts.
- Removed lmdb-dump ELF binary (superseded by lmdb-ffi.ts FFI).
